Improve portability.
accept4() is available on not just Linux, but also modern BSDs,
so use it when it's available.  I can't find documentation
on Darwin, so it will use the slower accept() path for now.

Traditional accept() has different semantics where the CLOEXEC and
NONBLOCK flags are inherited from the listening socket, so make
sure those flags are set properly on that path.

Most setsockopt options that are used are portable, except
for IP_BIND_ADDRESS_NO_PORT.