Use lockless list for gc.
This is all ad hoc, and I'm not sure that forward progress is
guaranteed between worker threads when trying to add elements
to the gc queue, but contention here is not terribly high
so it should be fine.

The listener thread that performs joins and frees is
straightforward and should be correct even under high contention.